requirements.txt
# --- Backend & API Framework ---
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# --- Machine Learning & AI ---
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.2
h5py==3.10.0

# --- Image Processing ---
Pillow==10.1.0
opencv-python==4.8.1.78

# --- Utilities ---
requests==2.31.0
python-dotenv==1.0.0